renamed AuthenticationCard to Authentication
