+++ src/components/organism/__test__/Authentication.test.tsx
@@ -1,13 +1,13 @@
 import 'jest-dom/extend-expect';
 import React from 'react';
 import { cleanup, render } from '../../../helpers/test-utils';
-import AuthenticationCard, { AuthType } from '../AuthenticationCard';
+import Authentication, { AuthType } from '../Authentication/Authentication';
 
 afterEach(cleanup);
 
 test('Login form renders with correct headertext, aria-labels and placeholders', () => {
   const { getByTestId, getByLabelText } = render(
-    <AuthenticationCard type={AuthType.Login} />,
+    <Authentication type={AuthType.Login} />,
   );
 
   expect(getByTestId('authform-header')).toHaveTextContent(/^Sign in$/);
@@ -15,9 +15,12 @@ test('Login form renders with correct headertext, aria-labels and placeholders',
   expect(getByLabelText('password')).toHaveAttribute('aria-label', 'password');
 });
 
-test('Register form renders with correct headertext', () => {
-  const { getByTestId } = render(
-    <AuthenticationCard type={AuthType.Register} />,
+test('Register form renders with correct headertext, aria-labels and placeholders', () => {
+  const { getByTestId, getByLabelText } = render(
+    <Authentication type={AuthType.Register} />,
   );
+
   expect(getByTestId('authform-header')).toHaveTextContent(/^Sign up$/);
+  expect(getByLabelText('email')).toHaveAttribute('aria-label', 'email');
+  expect(getByLabelText('password')).toHaveAttribute('aria-label', 'password');
 });
