+++ src/stories/Authentication.stories.tsx
@@ -2,14 +2,14 @@ import {storiesOf} from '@storybook/react';
 import {themes} from '@storybook/theming';
 import React from 'react';
 
-import AuthenticationCard, { AuthType } from '../components/molecules/AuthenticationCard';
+import Authentication, { AuthType } from '../components/organism/Authentication/Authentication';
 
 //  const Wrapper = styled.div`
 //   min-width: 200px;
 //   max-width: 400px;
 // `
 
-storiesOf('AuthenticationCard', module)
+storiesOf('Authentication', module)
   .addParameters({ options: { theme: themes.dark } })
-  .add('Login', () => <AuthenticationCard type={AuthType.Login} />)
-  .add('Register', () => <AuthenticationCard type={AuthType.Register} />);
+  .add('Login', () => <Authentication type={AuthType.Login} />)
+  .add('Register', () => <Authentication type={AuthType.Register} />);
